Move `Scheduler` and `ScheduledTask` to `snuba.utils.scheduler` to avoid circular import.
