+++ snuba/subscriptions/scheduler.py
@@ -1,48 +1,12 @@
 import math
-from abc import ABC, abstractmethod
-from dataclasses import dataclass
 from datetime import datetime, timedelta
-from typing import Generic, Iterator, List, Optional, TypeVar
+from typing import Iterator, List, Optional
 
 from snuba.subscriptions.data import PartitionId, Subscription, SubscriptionIdentifier
 from snuba.subscriptions.store import SubscriptionDataStore
 from snuba.utils.metrics.backends.abstract import MetricsBackend
+from snuba.utils.scheduler import Scheduler, ScheduledTask
 from snuba.utils.types import Interval
-
-
-TTask = TypeVar("TTask")
-
-
-@dataclass(frozen=True)
-class ScheduledTask(Generic[TTask]):
-    """
-    A scheduled task represents a unit of work (a task) that is intended to
-    be executed at (or around) a specific time.
-    """
-
-    # The time that this task was scheduled to execute.
-    timestamp: datetime
-
-    # The task that should be executed.
-    task: TTask
-
-
-class Scheduler(ABC, Generic[TTask]):
-    """
-    The scheduler maintains the scheduling state for various tasks and
-    provides the ability to query the schedule to find tasks that were
-    scheduled between a time interval.
-    """
-
-    @abstractmethod
-    def find(self, interval: Interval[datetime]) -> Iterator[ScheduledTask[TTask]]:
-        """
-        Find all of the tasks that were scheduled to be executed between the
-        lower bound (exclusive) and upper bound (inclusive) of the provided
-        interval. The tasks returned should be ordered by timestamp in
-        ascending order.
-        """
-        raise NotImplementedError
 
 
 class SubscriptionScheduler(Scheduler[Subscription]):
