+++ tests/subscriptions/test_scheduler.py
@@ -9,9 +9,10 @@ from snuba.subscriptions.data import (
     SubscriptionData,
     SubscriptionIdentifier,
 )
-from snuba.subscriptions.scheduler import ScheduledTask, SubscriptionScheduler
+from snuba.subscriptions.scheduler import SubscriptionScheduler
 from snuba.subscriptions.store import RedisSubscriptionDataStore
 from snuba.utils.metrics.backends.dummy import DummyMetricsBackend
+from snuba.utils.scheduler import ScheduledTask
 from snuba.utils.types import Interval
 from tests.base import BaseTest
